fix: resolvendo erro no download do arquivo

+++ app/Helpers/functions.php
@@ -4,20 +4,21 @@ use Illuminate\Support\Facades\Storage;
 
 function deleteDir($dir)
 {
-    if (Storage::exists($dir)) {
-        Storage::deleteDirectory($dir);
+    if (Storage::disk('public')->exists($dir)) {
+        Storage::disk('public')->deleteDirectory($dir);
     }
 }
 
 function upload_as($dir, $file, $name)
 {
-    if ($file != null) {
+    if ($file != null && is_file($file)) {
         deleteDir($dir);
-        if (is_file($file)) {
-            $upload = $file->storeAs($dir, $name . '.' . $file->extension());
-            return $upload;
-        }
-        return null;
+        $filename = $name . '.' . $file->extension();
+
+        // Salva no disco 'public' => storage/app/public/
+        $upload = $file->storeAs($dir, $filename, 'public');
+
+        return $upload; // retorna "documentos/arquivo.pdf"
     }
     return null;
 }
